+++ CHANGELOG.md
@@ -5,8 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `cluster instance deploy`: add profiles for astarte v1.2.
+
 ### Changed
 - Use Go 1.21 for releases.
+
+### Removed
+- `cluster instance deploy`: remove profiles for Astarte v0.10 and v0.11.
 
 ## [23.5.0] - 2023-10-04
 ### Added